refactor: Destroy DeployGateTaskFactory
