--- src/main/groovy/com/deploygate/gradle/plugins/tasks/factory/UploadArtifactTaskFactory.groovy
@@ -1,9 +1,0 @@
-package com.deploygate.gradle.plugins.tasks.factory
-
-import javax.annotation.Nonnull
-
-interface UploadArtifactTaskFactory<T> {
-    void registerUploadArtifactTask(@Nonnull T variantOrVariantNameOrCustomName, Object... dependsOn)
-
-    void registerAggregatedUploadArtifactTask(Object... dependsOn)
-}
